requirements.txt
markovify==0.7.1
pyyaml==3.12
tweepy==3.6.0
